requirements.txt
fastapi
jinja2
mysql-connector-python
tailwindcss
daisyui@latest